// src/types/navigation.ts
export type RootStackParamList = {
  Login: undefined
  Register: undefined
  ConfirmEmail: { email: string }
  KYCStatus: undefined
  MainTabs: undefined
  PendingApproval: undefined
  ForgotPassword: undefined
  ResetPassword: undefined
}

export type BottomTabParamList = {
  Home: undefined
  Explore: undefined
  Settings: undefined
}